+++ torch_contour/__init__.py
@@ -66,4 +66,4 @@ __all__ = [
     "CleanContours",
 ]
 
-__version__ = "1.4.4"
+__version__ = "1.4.5"